backend/requirements.txt
firebase-admin>=6.0.0
flask>=3.0.0
flask-cors>=4.0.0
twilio>=9.0.0
python-dotenv>=1.0.0
